Straighten backends and separate code generation code
