--- ffc/backends/ffc/__init__.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (C) 2011-2017 Martin Sandve Alnæs
-#
-# This file is part of FFC (https://www.fenicsproject.org)
-#
-# SPDX-License-Identifier:    LGPL-3.0-or-later
-"""The FFC specific backend to the UFLACS form compiler algorithms."""
