+++ ffc/codegeneration/evalderivs.py
@@ -5,7 +5,7 @@ import logging
 
 import numpy
 
-from ffc.backends.ufc.evaluatebasis import (generate_compute_basisvalues,
+from ffc.codegeneration.evaluatebasis import (generate_compute_basisvalues,
                                             generate_expansion_coefficients)
 
 logger = logging.getLogger(__name__)
