+++ ffc/codegeneration/finite_element.py
@@ -11,12 +11,12 @@
 
 from collections import defaultdict
 
-import ffc.backends.ufc.finite_element_template as ufc_finite_element
-from ffc.backends.ufc.evalderivs import (_generate_combinations,
+import ffc.codegeneration.finite_element_template as ufc_finite_element
+from ffc.codegeneration.evalderivs import (_generate_combinations,
                                          generate_evaluate_reference_basis_derivatives)
-from ffc.backends.ufc.evaluatebasis import generate_evaluate_reference_basis
-from ffc.backends.ufc.evaluatedof import generate_transform_values
-from ffc.backends.ufc.utils import (generate_return_int_switch,
+from ffc.codegeneration.evaluatebasis import generate_evaluate_reference_basis
+from ffc.codegeneration.evaluatedof import generate_transform_values
+from ffc.codegeneration.utils import (generate_return_int_switch,
                                     generate_return_new_switch)
 from ufl import product
