+++ ffc/uflacs/uflacsgenerator.py
@@ -9,7 +9,7 @@ source structure from factorized representation."""
 
 import logging
 
-from ffc.backends.ffc.backend import FFCBackend
+from ffc.codegeneration.backend import FFCBackend
 from ffc.representationutils import initialize_integral_code
 from ffc.uflacs.integralgenerator import IntegralGenerator
 from ffc.language.format_lines import format_indented_lines
